[3054] Cleanup - remove cause/workaround for duplicate subjects_ids
* No consequence for end users.
* Fixed because it was making it harder to reason about and write a fix
  for the bug at hand.

This patch does not cause `new.erb` to lose hidden fields because it
renders `_new_fields_holder.html.erb` which in turn renders all the
hidden fields for the params, it just removes the duplication of
subjects_ids hidden field(s).

Moving the rendering of `shared/course_creation_hidden_fields` out a
level for edit means edit is unaffected.