[3678] - Course vacancies all open
- When a user updates a course listing we want to trigger the vacancies
updated notification. The notification request is triggered when a
course has a single location and is updated from closed to open. The
notification is also triggered when a course has multiple locations and
all locations are updated from closed to open
- Extra test coverage added for the mulitple location scenario mentioned
above
- the '#post_request' method in the Base class has been updated to take
a 'body' parameter. This is so that we can indicate whether course
vacancies have been filled (true or false)